wip: refactor to koa server and fix following tests
all the tests pass now
