+++ package.json
@@ -22,14 +22,11 @@
     "knip": "^5.46.0",
     "msw": "^2.7.3",
     "prettier": "^3.5.3",
-    "tsx": "^4.20.5",
     "typescript-eslint": "^8.27.0",
     "vitest": "^3.0.9"
   },
   "dependencies": {
-    "@deno/kv": "^0.12.0",
     "@fedify/fedify": "^1.8.8",
-    "@hono/node-server": "^1.19.1",
     "@koa/bodyparser": "^5.1.1",
     "@koa/cors": "^5.0.0",
     "@koa/router": "^13.1.0",
