+++ src/test/setup.ts
@@ -1,12 +1,11 @@
-import { ServerType } from '@hono/node-server'
 import { getLogger } from '@logtape/logtape'
 import * as css from '@solid/community-server'
 import * as dotenv from 'dotenv'
+import { IncomingMessage, Server, ServerResponse } from 'http'
 import { dirname } from 'node:path'
 import { fileURLToPath } from 'node:url'
 import { afterAll, beforeAll, beforeEach } from 'vitest'
-import { AppConfig } from '../app.js'
-import { start } from '../server.js'
+import { AppConfig, createApp } from '../app.js'
 import { createRandomAccount, getRandomPort } from './helpers/index.js'
 import type { Person } from './helpers/types.js'
 
@@ -21,7 +20,7 @@ const appConfig: AppConfig = {
   baseUrl: '',
 }
 
-let server: ServerType
+let server: Server<typeof IncomingMessage, typeof ServerResponse>
 let person: Person
 let person2: Person
 let cssServer: css.App
@@ -34,8 +33,11 @@ beforeAll(() => {
   testConfig.cssPort = getRandomPort()
   testConfig.cssUrl = `http://localhost:${testConfig.cssPort}`
 
+  // appConfig.indexedGroups = [testConfig.cssUrl + '/group/group#us']
+  // appConfig.allowedGroups = appConfig.indexedGroups
   appConfig.port = getRandomPort()
   appConfig.baseUrl = `http://localhost:${appConfig.port}`
+  // appConfig.webId = new URL('/profile/card#bot', appConfig.baseUrl).toString()
 })
 
 beforeAll(async () => {
@@ -71,7 +73,13 @@ afterAll(async () => {
 })
 
 beforeAll(async () => {
-  server = start(appConfig)
+  const app = await createApp(appConfig)
+
+  server = await new Promise(resolve => {
+    const srv = app.listen(appConfig.port, () => {
+      resolve(srv)
+    })
+  })
 })
 
 afterAll(async () => {
